Initial Tweaks
Removed HTML-linked guides. That was a requirement of the project, which
I feel were wholly unnecessary.

Replaced FLA's trace() calls which, with hindsight, should have been
throwing errors.
Other miscellaneous changes and refactoring within the .fla and beyond.
Faded the animated flags in the background to 25% alpha, as they were
too distracting.
Removed mentionings of game types that weren't completed.